requirements.txt
streamlit
pypdf
python-dotenv
openai
langchain_community
langchain
chromadb
bs4
tiktoken